my commit

+++ webserver/thread.c
@@ -43,7 +43,7 @@ void *thread_main(void *arg)
 	struct conndata * cdata;
 
 	thread_data->thread_tid = pthread_self() / 256;
-	printf("Thread created [%lx]\n", (unsigned int) thread_data->thread_tid);
+	printf("Thread created [%lx]\n", (long unsigned int) thread_data->thread_tid);
 	fflush(stdout);
 
 	while(1) {
@@ -86,7 +86,7 @@ void *thread_main(void *arg)
 					Free(cdata);
 					close(cdata->socketint);
 					break;
-					}
+				}
 			}
 	}
